requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
attrs==23.2.0
    # via
    #   automat
    #   service-identity
    #   twisted
automat==22.10.0
    # via twisted
certifi==2024.2.2
    # via requests
cffi==1.16.0
    # via cryptography
charset-normalizer==3.3.2
    # via requests
constantly==23.10.4
    # via twisted
cryptography==42.0.2
    # via
    #   pyopenssl
    #   scrapy
    #   service-identity
cssselect==1.2.0
    # via
    #   parsel
    #   scrapy
filelock==3.13.1
    # via tldextract
hyperlink==21.0.0
    # via twisted
idna==3.6
    # via
    #   hyperlink
    #   requests
    #   tldextract
incremental==22.10.0
    # via twisted
itemadapter==0.8.0
    # via
    #   itemloaders
    #   scrapy
itemloaders==1.1.0
    # via scrapy
jmespath==1.0.1
    # via
    #   itemloaders
    #   parsel
lxml==5.1.0
    # via
    #   parsel
    #   scrapy
packaging==23.2
    # via
    #   parsel
    #   scrapy
parsel==1.8.1
    # via
    #   itemloaders
    #   scrapy
protego==0.3.0
    # via scrapy
pyasn1==0.5.1
    # via
    #   pyasn1-modules
    #   service-identity
pyasn1-modules==0.3.0
    # via service-identity
pycparser==2.21
    # via cffi
pydispatcher==2.0.7
    # via scrapy
pyopenssl==24.0.0
    # via scrapy
pyyaml==6.0.1
    # via -r requirements.in
queuelib==1.6.2
    # via scrapy
requests==2.31.0
    # via
    #   -r requirements.in
    #   requests-file
    #   tldextract
requests-file==2.0.0
    # via tldextract
scrapy==2.11.0
    # via -r requirements.in
service-identity==24.1.0
    # via scrapy
six==1.16.0
    # via automat
tldextract==5.1.1
    # via scrapy
twisted==22.10.0
    # via scrapy
typing-extensions==4.9.0
    # via twisted
urllib3==2.2.0
    # via requests
w3lib==2.1.2
    # via
    #   itemloaders
    #   parsel
    #   scrapy
zope-interface==6.1
    # via
    #   scrapy
    #   twisted

# The following packages are considered to be unsafe in a requirements file:
# setuptools
